move dicodile/data/home.py to dicodile/_config.py and make necessary import updates
