+++ benchmarks/odl_text_runtime.py
@@ -10,6 +10,7 @@ from dicodile.utils import check_random_state
 import spams
 from sklearn.feature_extraction.image import extract_patches_2d
 
+from dicodile._config import DATA_HOME
 # Import to initiate the dictionary
 from dicodile.utils.dictionary import prox_d
 from dicodile.update_d.update_d import tukey_window
@@ -24,7 +25,7 @@ memory = Memory(location='.', verbose=0)
 
 BASE_FILE_NAME = Path(__file__).with_suffix('').name
 OUTPUT_DIR = Path('benchmarks_results')
-DATA_DIR = Path('../..') / 'data' / 'images' / 'text'
+DATA_DIR = DATA_HOME / 'images' / 'text'
 
 
 @memory.cache
