+++ dicodile/data/images.py
@@ -2,8 +2,7 @@ import PIL
 import matplotlib.pyplot as plt
 from download import download
 
-from .home import DATA_HOME
-
+from dicodile._config import DATA_HOME
 
 def fetch_mandrill():
